♻️ refactor(hcaptcha): rename client_response to token module
- rename client_response.rs to token.rs for clearer module naming
- update Token import to reflect module name change
