+++ crates/captval/src/hcaptcha.rs
@@ -1,5 +1,3 @@
-#[doc(hidden)]
-mod client_response;
 mod code;
 #[cfg(not(feature = "ext"))]
 #[doc(hidden)]
@@ -9,11 +7,12 @@ mod secret;
 mod secret_ext;
 #[doc(hidden)]
 mod sitekey;
+#[doc(hidden)]
+mod token;
 
 // pub(crate) use client_response::ClientResponse;
 pub use code::Code;
 
-pub use client_response::Token;
 #[cfg(not(feature = "ext"))]
 pub(crate) use secret::Secret;
 #[allow(unused_imports)]
@@ -21,6 +20,7 @@ pub(crate) use secret::Secret;
 pub(crate) use secret_ext::Secret;
 #[allow(unused_imports)]
 pub(crate) use sitekey::Sitekey;
+pub use token::Token;
 
 /// Endpoint url for the Hcaptcha siteverify API.
 pub const VERIFY_URL: &str = "https://api.hcaptcha.com/siteverify";
